+++ src/trusted/validator_x86/ncenuminsts_x86_32.c
@@ -96,7 +96,7 @@ Bool NaClInstValidates(uint8_t* mbase,
 }
 
 Bool NaClSegmentValidates(uint8_t* mbase,
-                          uint8_t size,
+                          size_t size,
                           NaClPcAddress vbase) {
   NaClCPUFeaturesX86 cpu_features;
   NaClValidationStatus  status;
